feat: use Spring Repository
refs #7
