+++ service/src/main/java/su/itpro/tasktracker/repository/AccountFilterRepository.java
@@ -1,0 +1,11 @@
+package su.itpro.tasktracker.repository;
+
+import java.util.Optional;
+import su.itpro.tasktracker.model.dto.AccountLoginDto;
+import su.itpro.tasktracker.model.entity.Account;
+
+public interface AccountFilterRepository {
+
+  Optional<Account> findByLoginOrEmail(AccountLoginDto loginDto);
+
+}
