+++ service/src/main/java/su/itpro/tasktracker/repository/CategoryRepository.java
@@ -1,24 +1,10 @@
 package su.itpro.tasktracker.repository;
 
-import static su.itpro.tasktracker.model.entity.QCategory.category;
-
-import com.querydsl.jpa.impl.JPAQuery;
-import jakarta.persistence.EntityManager;
-import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import su.itpro.tasktracker.model.entity.Category;
 
 @Repository
-public class CategoryRepository extends BaseRepository<Integer, Category> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 
-  public CategoryRepository(EntityManager entityManager) {
-    super(entityManager, Category.class);
-  }
-
-  public List<Category> findAll() {
-    return new JPAQuery<Category>(entityManager)
-        .select(category)
-        .from(category)
-        .fetch();
-  }
 }
