+++ service/src/main/java/su/itpro/tasktracker/repository/GroupRepository.java
@@ -1,15 +1,12 @@
 package su.itpro.tasktracker.repository;
 
-import jakarta.persistence.EntityManager;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import su.itpro.tasktracker.model.entity.Group;
 
 @Repository
-public class GroupRepository extends BaseRepository<UUID, Group> {
+public interface GroupRepository extends JpaRepository<Group, UUID> {
 
-  public GroupRepository(EntityManager entityManager) {
-    super(entityManager, Group.class);
-  }
 
 }
