+++ service/src/main/java/su/itpro/tasktracker/repository/ProfileRepository.java
@@ -1,15 +1,12 @@
 package su.itpro.tasktracker.repository;
 
-import jakarta.persistence.EntityManager;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import su.itpro.tasktracker.model.entity.Profile;
 
 @Repository
-public class ProfileRepository extends BaseRepository<UUID, Profile> {
+public interface ProfileRepository extends JpaRepository<Profile, UUID> {
 
-  public ProfileRepository(EntityManager entityManager) {
-    super(entityManager, Profile.class);
-  }
 
 }
