+++ service/src/main/java/su/itpro/tasktracker/repository/TaskFilterRepository.java
@@ -1,0 +1,11 @@
+package su.itpro.tasktracker.repository;
+
+import java.util.List;
+import su.itpro.tasktracker.model.dto.TaskFilter;
+import su.itpro.tasktracker.model.entity.Task;
+
+public interface TaskFilterRepository {
+
+  List<Task> findAllByFilter(TaskFilter filter);
+
+}
