+++ service/src/main/resources/application.yml
@@ -13,8 +13,7 @@ spring:
       hibernate:
         show_sql: true
         format_sql: true
-        hbm2ddl:
-          auto: validate
+        hbm2ddl.auto: validate
 logging:
   level:
     root: info
