+++ service/src/main/resources/db/changelog/db.changelog-1.0.sql
@@ -1,16 +1,23 @@
+--liquibase formatted sql
+
+--changeset s.ivaschenko:1
 CREATE TABLE category
 (
     id SERIAL PRIMARY KEY ,
     name VARCHAR(64) NOT NULL UNIQUE ,
     CONSTRAINT check_name_con CHECK ( name <> '' )
 );
+--rollback DROP TABLE category;
 
+--changeset s.ivaschenko:2
 CREATE TABLE groups
 (
     id UUID PRIMARY KEY ,
     name VARCHAR(64) NOT NULL UNIQUE
 );
+--rollback DROP TABLE groups;
 
+--changeset s.ivaschenko:3
 CREATE TABLE account
 (
     id UUID PRIMARY KEY ,
@@ -21,7 +28,9 @@ CREATE TABLE account
     role VARCHAR(8) NOT NULL ,
     CONSTRAINT check_role_con CHECK ( role IN ('USER', 'ADMIN') )
 );
+--rollback DROP TABLE account;
 
+--changeset s.ivaschenko:4
 CREATE TABLE profile
 (
     account_id UUID PRIMARY KEY REFERENCES account (id),
@@ -31,14 +40,18 @@ CREATE TABLE profile
     gender VARCHAR(6) ,
     CONSTRAINT check_gender_con CHECK ( gender IN ('MALE', 'FEMALE') )
 );
+--rollback DROP TABLE profile;
 
+--changeset s.ivaschenko:5
 CREATE TABLE project
 (
     id UUID PRIMARY KEY ,
     name VARCHAR(128) NOT NULL UNIQUE ,
     description text
 );
+--rollback DROP TABLE project;
 
+--changeset s.ivaschenko:6
 CREATE TABLE task
 (
     id BIGSERIAL PRIMARY KEY ,
@@ -56,5 +69,4 @@ CREATE TABLE task
     description TEXT ,
     CONSTRAINT check_title_con CHECK ( task.title <> '' )
 )
-
-
+--rollback DROP TABLE task;
