+++ service/src/test/resources/application-test.yml
@@ -4,4 +4,4 @@ spring:
     password: test
   jpa:
     properties.hibernate:
-      hbm2ddl.auto: update
+      hbm2ddl.auto: validate
